pyproject.toml
[project]
name = "owrestimator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "altair==5.4.1",
    "attrs==24.2.0",
    "beautifulsoup4==4.12.3",
    "blinker==1.9.0",
    "cachetools==5.5.0",
    "certifi==2024.8.30",
    "charset-normalizer==3.4.0",
    "click==8.1.7",
    "contourpy==1.3.0",
    "cycler==0.12.1",
    "flask==3.0.3",
    "fonttools==4.54.1",
    "gitdb==4.0.11",
    "gitpython==3.1.43",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.4",
    "joblib==1.1.0",
    "jsonschema-specifications==2024.10.1",
    "jsonschema==4.23.0",
    "kiwisolver==1.4.7",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.9.2",
    "mdurl==0.1.2",
    "narwhals==1.13.3",
    "numpy==1.26.4",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==11.0.0",
    "protobuf==5.28.3",
    "pyarrow==18.0.0",
    "pydeck==0.9.1",
    "pygments==2.18.0",
    "pyparsing==3.2.0",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.2",
    "referencing==0.35.1",
    "requests==2.32.3",
    "rich==13.9.4",
    "rpds-py==0.21.0",
    "scikit-learn==1.0.2",
    "scipy==1.14.1",
    "six==1.16.0",
    "sklearn-pandas==2.2.0",
    "smmap==5.0.1",
    "soupsieve==2.6",
    "srcomapi==0.3.3",
    "streamlit==1.40.0",
    "tenacity==9.0.0",
    "threadpoolctl==3.5.0",
    "toml==0.10.2",
    "tornado==6.4.1",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "urllib3==2.2.3",
    "watchdog==5.0.3",
    "werkzeug==3.1.3",
]

[dependency-groups]
lint = [
    "ruff>=0.14.10",
]
dev = [
    "pytest-watcher>=0.6.2",
]
test = [
    "pytest-cov>=7.0.0",
    "pytest>=9.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["owrestimator"]